refactor: better logging in the browser
