+++ packages/core/src/index.ts
@@ -7,7 +7,6 @@ export * from "./log/Controller.definitions.js";
 export * from "./log/Controller.js";
 export * from "./log/ZWaveLoggerBase.js";
 export * from "./log/shared.js";
-export * from "./log/shared_safe.js";
 export type * from "./log/traits.js";
 export * from "./qr/index.js";
 export * from "./reflection/decorators.js";
