+++ packages/core/src/log/Colorizer.ts
@@ -1,13 +1,8 @@
 import colors from "ansi-colors";
-import { type TransformFunction, format } from "logform";
-import winston from "winston";
-import type { ZWaveLogInfo } from "./shared_safe.js";
-const defaultColors = winston.config.npm.colors;
+import { configs } from "triple-beam";
+import { type LogFormat } from "./format.js";
 
-// This is a placeholder
-interface ColorizerOptions {
-	__foo?: undefined;
-}
+const defaultColors = configs.npm.colors;
 
 const primaryAndInlineTagRegex = /\[([^\]]+)\]/g;
 
@@ -29,44 +24,45 @@ function colorizeTextAndTags(
 	);
 }
 
-export const colorizer = format(
-	((
-		info: ZWaveLogInfo,
-		_opts: ColorizerOptions,
-	) => {
-		const levelColorKey =
-			defaultColors[info.level as keyof typeof defaultColors] as string;
-		const textColor = (colors as any)[levelColorKey];
-		const bgColor = (colors as any)[getBgColorName(levelColorKey)];
-		// Colorize all segments separately
-		if (typeof info.message === "string") {
-			info.message = colorizeTextAndTags(
-				info.message,
-				textColor,
-				bgColor,
-			);
-		} else {
-			info.message = info.message.map((msg) =>
-				colorizeTextAndTags(msg, textColor, bgColor)
-			);
-		}
-		info.direction = colors.white(info.direction);
-		if (info.label) {
-			info.label = colors.gray.inverse(info.label);
-		}
-		if (info.timestamp) {
-			info.timestamp = colors.gray(info.timestamp);
-		}
-		if (info.primaryTags) {
-			info.primaryTags = colorizeTextAndTags(
-				info.primaryTags,
-				textColor,
-				bgColor,
-			);
-		}
-		if (info.secondaryTags) {
-			info.secondaryTags = colors.gray(info.secondaryTags);
-		}
-		return info;
-	}) as unknown as TransformFunction,
-);
+export function colorizer(bg: boolean = true): LogFormat {
+	return {
+		transform: (info) => {
+			const levelColorKey =
+				defaultColors[info.level as keyof typeof defaultColors];
+			const textColor = (colors as any)[levelColorKey];
+			const bgColor = bg
+				? (colors as any)[getBgColorName(levelColorKey)]
+				: ((txt: string) => txt);
+			// Colorize all segments separately
+			if (typeof info.message === "string") {
+				info.message = colorizeTextAndTags(
+					info.message,
+					textColor,
+					bgColor,
+				);
+			} else {
+				info.message = info.message.map((msg) =>
+					colorizeTextAndTags(msg, textColor, bgColor)
+				);
+			}
+			info.direction = colors.white(info.direction);
+			if (info.label) {
+				info.label = colors.gray.inverse(info.label);
+			}
+			if (info.timestamp) {
+				info.timestamp = colors.gray(info.timestamp);
+			}
+			if (info.primaryTags) {
+				info.primaryTags = colorizeTextAndTags(
+					info.primaryTags,
+					textColor,
+					bgColor,
+				);
+			}
+			if (info.secondaryTags) {
+				info.secondaryTags = colors.gray(info.secondaryTags);
+			}
+			return info;
+		},
+	};
+}
